requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-mysql==3.9.0
django-sendgrid-v5==0.9.0
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
jdcal==1.4.1
oauthlib==3.1.0
openpyxl==3.0.5
Pillow==8.0.1
psycopg2==2.7.5
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==1.5
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.4.7
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
urllib3==1.25.11
whitenoise==5.1.0
XlsxWriter==1.3.7
xlwt==1.3.0
